feat(schema-fetch): add support for fetching schemas. refactoring
