+++ packages/utils/package.json
@@ -21,6 +21,7 @@
         "@semantic-release/release-notes-generator": "^7.1.4",
         "dhis2-uid": "^0.1.2",
         "ejs": "^2.6.1",
+        "inquirer": "^6.3.1",
         "jsondiffpatch": "^0.3.11",
         "semantic-release": "^15.13.3"
     },
